feat(sneaker_views_clickhouse_writer): Changed handle handle_sneaker_view_to_click house The functionality used to be: She called a function that made an entry in the sneaker_views_histories table in ClickHouse with the data that was passed to her.
Updated functionality:
Initially, a ClickHouse request is made to the sneaker_views_histories table in order to check if there is already a record with a specific user_id and sneaker_id in it.

Then there are 2 scenarios:

1) The record was not in the table
Since there is no record in the table, we can simply create it, call the clickhouse_insert function and create a record with the necessary data.

2) The record ended up in the table
First of all, an entry is created in ClickHouse with the same data as the previous entry, only the value version = version + 1 and sign = -1. Why we do this -> read the full description of the feature.
The second step is to create an entry in ClickHouse, but with new data, that is, a new view_timestamp and the values of the required fields: sign = 1, version = version + 1.