feat(cart_service): Added calculation of the total amount for the basket
The handle that displays the user's shopping cart has added the logic of calculating the total amount for all the items in it.
The functionality requires improvements, such as taking the logic of calculating the amount into a separate function, writing a response_model to return only the necessary data to the client.

This functionality was added due to the fact that the PayPal payment system will be connected in the near future.